requirements.txt
opencv-python
numpy
pillow
ultralyticsplus
iqoptionapi @ git+https://github.com/iqoptionapi/iqoptionapi.git